Integration of the service container into ZF dispatching process.
LoSo_Zend_Controller_Dispatcher_SymfonyContainerDispatcher is set into the FrontController instead of ZF standard dispatcher and can then load controllers from the service container when necessary instead of instantiating them.
As a consequence, the service container is integrated into the heart of ZF and is responsible for loading controllers with dependency management.

To increase performances in production when loading XML, YAML or INI configuration files or when reflecting annotated classes, the container can be cached once completely loaded.